VigésimoPrimero commit. BackEnd_02

+++ app/src/main/java/com/arq_control/ui/visitas/OnNuevaVisitaListener.java
@@ -7,7 +7,7 @@ public interface OnNuevaVisitaListener {
                                             long numeroVisita,
                                             String descripcion);
 
-   public void onVisitaActualizarClickListener(int id,
+   public void onVisitaActualizarClickListener(long id,
                                                 String tituloVisita,
                                                 String fecha,
                                                 long numeroVisita,
